+++ sql/update_to_1.4.0.sql
@@ -1,0 +1,4 @@
+CREATE TABLE user_settings (
+    id INTEGER PRIMARY KEY,
+    setting TEXT
+);